// src/data/action-type/pbsupl-action-type.ts
enum PbSuplActionType {
    PB_SUP_GET_ALL = "PB_SUP_GET_ALL",
    PB_SUP_GET_OVERVIEW = "PB_SUP_GET_OVERVIEW",
    PB_SUP_CHECK_ROSSO = "PB_SUP_CHECK_ROSSO",
    PB_SUP_PLU_VALIDATION ="PB_SUP_PLU_VALIDATION",
    PB_SUP_ADD_OVERVIEW_DRAFT = "PB_SUP_ADD_OVERVIEW_DRAFT",
    PB_SUP_DELETE_OVERVIEW_DRAFT = "PB_SUP_DELETE_OVERVIEW_DRAFT",
    PB_SUP_EDIT_OVERVIEW_DRAFT = "PB_SUP_EDIT_OVERVIEW_DRAFT",
    PB_SUP_QTY_ORDER_VALIDATION = "PB_SUP_QTY_ORDER_VALIDATION",
    PB_SUP_CLEAR_ALL = "PB_SUP_CLEAR_ALL",
    PB_SUP_SAVE = "PB_SUP_SAVE",
    PB_SUP_VERIFY ="PB_SUP_VERIFY"
}

export {
    PbSuplActionType
};